re-factor install command
