+++ packages/tools/src/command.ts
@@ -3,6 +3,7 @@ import yargs from 'yargs';
 export interface Command {
 	synopsis(): string;
 	description(): string;
-	options(): { [key: string]: yargs.Options };
-	run(argv: yargs.Arguments): void;
+	args(): { [key: string]: yargs.Options };
+	init(argv: yargs.Arguments): void;
+	run(): Promise<number>;
 }
